Added internal data dump on verify, ignore connection errors

+++ parameters/base.go
@@ -122,7 +122,7 @@ type Types struct {
 
 // Arguments represents any arguments that are passed into the program at runtime.
 type Arguments struct {
-	NumOfCycles      int
+	NumOfCycles      int64
 	Timeout          time.Duration
 	FirstError       bool
 	ConfigPath       string
